requirements.txt
ultralytics
opencv-python
pillow
numpy
matplotlib
pytesseract
torch
torchvision
